[오류수정] 구조개선으로 인한 프리스타일 음악 재시작이 되지않던 오류 수정

+++ Assets/Scripts/IO/FileConverter.cs
@@ -30,7 +30,7 @@ public struct Song
     public string videoName;
     public int    audioOffset;
     public int    videoOffset;
-    public int    volume;
+    public float  volume;
 
     public int totalTime;
     public int previewTime;
